# docs-backup-20260213-150302/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ATO Tax Optimizer API
  version: 0.1.0
  description: >
    API for the Australian Tax Optimizer platform. Integrates with Xero for
    financial data, uses Gemini AI for forensic analysis, and provides tax
    optimisation recommendations for Australian businesses.
  contact:
    name: ATO Tax Optimizer
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://ato-app.vercel.app
    description: Production (Vercel)
security:
  - bearerAuth: []
...
paths:
...
  /api/share/{token}/feedback:
    get:
      tags: [share]
      summary: Get feedback on shared report
      description: Returns feedback/comments from report recipients.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Feedback list
    post:
      tags: [share]
      summary: Submit feedback on shared report
      description: Allows a report recipient to leave feedback.
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Feedback submitted
...
  /api/rnd/registrations:
    get:
      tags: [rnd]
      summary: List R&D registrations
      description: Returns R&D activity registrations for a tenant.
      parameters:
        - $ref: '#/components/parameters/tenantIdQuery'
      responses:
        '200':
          description: Registration list
    post:
      tags: [rnd]
      summary: Create R&D registration
      description: Creates a new R&D activity registration record.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId]
              properties:
                tenantId:
                  type: string
      responses:
        '201':
          description: Registration created
...
  /api/organizations/groups:
    get:
      tags: [organizations]
      summary: List organisation groups
      description: Returns all organisation groups for the current user.
      responses:
        '200':
          description: Group list
    post:
      tags: [organizations]
      summary: Create organisation group
      description: Creates a new group for consolidating organisations.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Group created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token from authentication
  parameters:
    tenantIdQuery:
      name: tenantId
      in: query
      required: true
      description: Xero tenant ID
      schema:
        type: string
...